+++ libs/ui/src/components/DisablableTab/DisablableTab.scss
